domains/brainwire/pyproject.toml
[project]
name = "brainwire"
version = "2.0.0"
description = "Universal Consciousness State Engine"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
bw-bench = "brainwire.bench:main"
bw-calc = "brainwire.calc:main"
